Drop expr module
